+++ specs/dataframe/serialization_spec.lua
@@ -1,22 +1,30 @@
 require 'lfs'
 
--- Make sure that directory structure is always the same
-if (string.match(lfs.currentdir(), "/specs$")) then
+-- Ensure the test is launched within the specs/ folder
+assert(string.match(lfs.currentdir(), "specs")~=nil, "You must run this test in specs folder")
+
+local initial_dir = lfs.currentdir()
+
+-- Go to specs folder
+while (not string.match(lfs.currentdir(), "/specs$")) do
   lfs.chdir("..")
 end
 
--- Include Dataframe lib
-dofile('init.lua')
+local specs_dir = lfs.currentdir()
+lfs.chdir("..")-- one more directory and it is lib root
 
--- Go into specs so that the loading of CSV:s is the same as always
-lfs.chdir("specs")
+-- Include Dataframe lib
+dofile("init.lua")
+
+-- Go back into initial dir
+lfs.chdir(initial_dir)
 
 describe("Serialization", function()
 	-- Do not use advanced_short since it has nan that are 0/0 ~= 0/0 == true
 	local df = Dataframe()
 
 	it("Deserializes a simple Dataframe object",function()
-		df:load_csv{path = "./data/simple_short.csv", verbose = false}
+		df:load_csv{path = specs_dir.."/data/simple_short.csv", verbose = false}
 
 		b = torch.serialize(df)
 		c = torch.deserialize(b)
@@ -38,7 +46,7 @@ describe("Serialization", function()
 	end)
 
 	it("Saves with init",function()
-		local a = Dataframe("./data/realistic_29_row_data.csv")
+		local a = Dataframe(specs_dir.."/data/realistic_29_row_data.csv")
 
 		a:create_subsets()
 		a:fill_all_na()